blogging system done

+++ app/Http/Controllers/PostsController.php
@@ -1,13 +1,23 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use Illuminate\Http\Request;
+use  Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -43,12 +53,32 @@ class PostsController extends Controller
     {
         $this->validate($request,[
             'title'=>'required',
-            'body'=>'required'
+            'body'=>'required',
+            'cover_image'=>'image|nullable|max:1999'
         ]);
+//Handle File Upload
+if($request->hasFile('cover_image')){
+   //Get filename with the extension
+   $finenameWithExt=$request->file('cover_image')->getClientOriginalName();
+   //Get just filename
+    $filename=pathinfo($finenameWithExt,PATHINFO_FILENAME);
+   //Get just extension
+   $extension=$request->file('cover_image')->getClientOriginalExtension();
+
+   //Filename to store
+   $fileNameToStore=$filename.'_'.time().'.'.$extension;
+
+   //Upload Image
+   $path =$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+}else{
+    $fileNameToStore="noimage.jpg";
+}
+
        //Create Post
        $post=new Post;
        $post->title=$request->input('title');
        $post->body=$request->input('body');
+       $post->cover_image=$fileNameToStore;
        $post->user_id=auth()->user()->id;
        $post->save();
 
@@ -76,6 +106,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post=  Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !==$post->user_id){
+        return redirect('/posts')->with('error','Unauthorized page');
+        }
         return view('posts.edit')->with('post',$post);
 
 
@@ -94,10 +129,30 @@ class PostsController extends Controller
             'title'=>'required',
             'body'=>'required'
         ]);
+
+        //Handle File Upload
+if($request->hasFile('cover_image')){
+    //Get filename with the extension
+    $finenameWithExt=$request->file('cover_image')->getClientOriginalName();
+    //Get just filename
+     $filename=pathinfo($finenameWithExt,PATHINFO_FILENAME);
+    //Get just extension
+    $extension=$request->file('cover_image')->getClientOriginalExtension();
+ 
+    //Filename to store
+    $fileNameToStore=$filename.'_'.time().'.'.$extension;
+ 
+    //Upload Image
+    $path =$request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+ }
+ 
        //Update Post
        $post=Post::find($id);
        $post->title=$request->input('title');
        $post->body=$request->input('body');
+       if($request->hasFile('cover_image')){
+           $post->cover_image= $fileNameToStore;
+       }
        $post->save();
 
        return redirect('/posts')->with('success','Post Updated');
@@ -112,8 +167,20 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post=Post::find($id);
-        $post->delete();
-        return redirect('/posts')->with('success','Post Deleted');
+         //Check for correct user
+         if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized page');
+            }
+         
+            if($post->cover_image !='noimage.jpg'){
+                //Delete image
+                 Storage::delete('public/cover_images/'.$post->cover_image);
+            }
+
+            
+             $post->delete();
+             return redirect('/posts')->with('success','Post Deleted');
 
     }
+
 }
